Handled files inconsistencies
Handled file inconsistencies to check where the time was even if it was a different character used. Also now handling for capital letters with more if statements. Also made changes to the "rest of the day" case to make sure we're working out that time correctly (was resetting a value inside the loop when shouldn't be)