Fixes
Minor fixes to problems with the loops not deleting  empty elements that occured after each other. Changing if to while loop fixed this. Also made sure we're deducting the time from the split job. This won't handle non double values for now but will make the changes in main code to convert all strings to double readable values.